+++ src/components/index.js
@@ -4,21 +4,13 @@ import FadePage from './FadePage';
 import LoadingBar from './Loading/LoadingBar';
 import LoadingCircle from './Loading/LoadingCircle';
 import ShowError from './ShowError';
-import Popoup from './popups/Popoup.jsx';
 import Card from './Cards/Card.jsx';
 import CardAdd from './Cards/CardAdd.jsx';
 import Loader from './Loading/Loader.jsx';
-import Dialog from './popups/Dialog.jsx';
 import DialogLib from './popups/DialogLib.jsx';
 import Checkbox from './Inputs/Checkbox.jsx';
-import ChangePass from './popups/ChangePass.jsx';
-import DialogImage from './popups/DialogImage.jsx';
 import Dropdown from './Dropdown.jsx';
-import DialogUpload from './popups/DialogUpload.jsx';
-import Logout from './popups/Logout.jsx';
 import Container from './Container.jsx';
-import AdminHeader from './Headers/AdminHeader.jsx';
-import AdminFooter from './Footers/AdminFooter.jsx';
 import Header from './Headers/Header.jsx';
 import Footer from './Footers/Footer.jsx';
 import PageNotfound from './PageNotfound.jsx';
@@ -31,21 +23,13 @@ export {
     LoadingBar,
     LoadingCircle,
     ShowError,
-    Popoup,
     Card,
     CardAdd,
     Loader,
-    Dialog,
     DialogLib,
     Checkbox,
-    ChangePass,
-    DialogImage,
     Dropdown,
-    DialogUpload,
-    Logout,
     Container,
-    AdminFooter,
-    AdminHeader,
     Header,
     Footer,
     PageNotfound,
