+++ src/layouts/Profile/Components/ChangePass.jsx
@@ -1,9 +1,9 @@
 import React, { Fragment, useEffect, useRef, useState } from 'react';
 import { Dialog, Transition } from '@headlessui/react';
 import { LockClosedIcon } from '@heroicons/react/24/outline';
-import { FadePage, Input, ShowError } from '../index.js';
-import useApi from '../../hooks/useApi.js';
-import { api } from '../../constants/index.js';
+import { FadePage, Input, ShowError } from '../../../components/index.js';
+import useApi from '../../../hooks/useApi.js';
+import { api } from '../../../constants/index.js';
 import LoadingBar from 'react-top-loading-bar';
 
 export default function ChangePass({ open, onClose, Heading = 'Heading' }) {
