+++ src/layouts/Profile/Components/Logout.jsx
@@ -2,8 +2,8 @@ import { Fragment, useEffect, useRef } from 'react';
 import { Dialog, Transition } from '@headlessui/react';
 import { ExclamationTriangleIcon } from '@heroicons/react/24/outline';
 import { useNavigate } from 'react-router-dom';
-import useApi from '../../hooks/useApi.js';
-import { api } from '../../constants/index.js';
+import useApi from '../../../hooks/useApi.js';
+import { api } from '../../../constants/index.js';
 
 export default function Logout({
     open,
