+++ src/layouts/Profile/Components/index.js
@@ -1,0 +1,6 @@
+import ChangePass from './ChangePass.jsx';
+import DialogImage from './DialogImage.jsx';
+import DialogUpload from './DialogUpload.jsx';
+import Logout from './Logout.jsx';
+
+export { ChangePass, DialogImage, DialogUpload, Logout };
